Change backups to breakup images into smaller files for github. Restore images in batches to keep memory use low.
Former-commit-id: 48d2af5bfd988bc7fbf56ae7f3d31660dc83c237

+++ lib/tasks/restore.rake
@@ -1,55 +1,85 @@
 DIR = "material/backup/"
+IMAGE_COLUMNS = [:id, :imgurId, :nsfw, :reddit_score, :gif, :reported, :deleted, :updated_at, :created_at]
+# If we try to insert all images at once it uses quite a bit of memory (> 1 gb).
+# We can insert in batches to keep memory use low
+IMAGE_INSERT_BATCH_SIZE = 25000
 
 namespace :restore do
  
   desc "Restore all images from json file"
   task images: :environment do
+    startTime = Time.now
+
     # Delete existing images and associated categories
     Image.delete_all
     # And clear their categories
     ActiveRecord::Base.connection.execute("delete from categories_images") 
-
-    imageColumns = [:id, :imgurId, :nsfw, :reddit_score, :gif, :reported, :deleted, :updated_at, :created_at]
+    
+    count = 0
     images = []
-    category_inserts = []
+    category_data = []
 
-    filepath = "#{DIR}images_backup.json"
-    File.open(filepath, 'r') do |f|
-      while (line = f.gets)
-        json = JSON.parse(line)
+    # Go through each of the files holding images. They are of the form "images_backup_#{file_number}"
+    imagesFilePath = "#{DIR}images_backup_"
+    imageFileCount = 0
+    
+    loop do
+      imageFileCount += 1
+      filepath = imagesFilePath + imageFileCount.to_s
+      break if !File.exists? filepath
       
-        id = json["id"]
-        imgurId = json["imgurId"]
-        nsfw = json["nsfw"]
-        categories = json["categories"]
-        reddit_score = json["reddit_score"]
-        gif = json["gif"]
-        reported = json["reported"]
-        deleted = json["deleted"]
-        updated_at = json["updated_at"]
-        created_at = json["created_at"]
-
-
-        images << [id, imgurId, nsfw, reddit_score, gif, reported, deleted, updated_at, created_at]
+      File.open(filepath, 'r') do |f|
+        while (line = f.gets)
+          json = JSON.parse(line)
         
-        categories.each do |category_id|
-          category_inserts << "(#{id}, #{category_id})"
-        end
+          id = json["id"]
+          imgurId = json["imgurId"]
+          nsfw = json["nsfw"]
+          categories = json["categories"]
+          reddit_score = json["reddit_score"]
+          gif = json["gif"]
+          reported = json["reported"]
+          deleted = json["deleted"]
+          updated_at = json["updated_at"]
+          created_at = json["created_at"]
+
+          images << [id, imgurId, nsfw, reddit_score, gif, reported, deleted, updated_at, created_at]
+          
+          categories.each do |category_id|
+            category_data << "(#{id}, #{category_id})"
+          end
+
+          count += 1
+
+          # If we have enough images then do the insert and clear the collected data
+          if count % IMAGE_INSERT_BATCH_SIZE == 0
+            insert_images(images, category_data)
+            images.clear
+            category_data.clear
+          end
+
+        end        
       end
     end
 
-    Image.import imageColumns, images, validate: false
+    # Insert the rest of the data
+    insert_images(images, category_data)
+    puts "#{count} images restored in #{Time.now - startTime} seconds"
+  end
+
+  # Efficiently mass insert image data into the database
+  def insert_images(images_array, categories_array)
+    # Import image data using ActiveRecord-Import gem. Provide model columns and array of images
+    Image.import IMAGE_COLUMNS, images_array, validate: false
 
     #Insert the category relationships with direct sql. Sqlite limits mass inserts to 500 at a time.
     # We don't want to pad the extra space in the last batch so pass false to fill_with
     ActiveRecord::Base.transaction do    
-      category_inserts.in_groups_of(450, false) do |batch| 
+      categories_array.in_groups_of(450, false) do |batch| 
         category_sql = "INSERT INTO categories_images (image_id, category_id) VALUES #{batch.join(", ")}"
         ActiveRecord::Base.connection.execute(category_sql) 
       end
     end
-
-    puts "#{images.size} images restored"
   end
 
   desc "Restore albums"
@@ -63,7 +93,7 @@ namespace :restore do
     albums = []
     category_inserts = []
 
-    filepath = "#{DIR}albums_backup.json"
+    filepath = "#{DIR}albums_backup"
     File.open(filepath, 'r') do |f|
       while (line = f.gets)
         json = JSON.parse(line)
@@ -107,7 +137,7 @@ namespace :restore do
 
     categories = []
 
-    filepath = "#{DIR}categories_backup.json"
+    filepath = "#{DIR}categories_backup"
     File.open(filepath, 'r') do |f|
       while (line = f.gets)
         json = JSON.parse(line)
